# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EnvironmentStatePathStatus = typeof EnvironmentStatePathStatus[keyof typeof EnvironmentStatePathStatus];


export const EnvironmentStatePathStatus = {
  clear: 'clear',
  blocked: 'blocked',
} as const;

export type EnvironmentStateElevator = typeof EnvironmentStateElevator[keyof typeof EnvironmentStateElevator];


export const EnvironmentStateElevator = {
  available: 'available',
  unavailable: 'unavailable',
  unknown: 'unknown',
} as const;

export type OcrResultDocumentType = typeof OcrResultDocumentType[keyof typeof OcrResultDocumentType];


export const OcrResultDocumentType = {
  sign: 'sign',
  document: 'document',
  unknown: 'unknown',
} as const;

export interface OcrResult {
  text: string;
  lines: string[];
  documentType: OcrResultDocumentType;
  confidence: number;
  demoMode: boolean;
  safetyMessage: string;
}

export type SpatialWaypointType = typeof SpatialWaypointType[keyof typeof SpatialWaypointType];


export const SpatialWaypointType = {
  origin: 'origin',
  turn: 'turn',
  destination: 'destination',
  elevator: 'elevator',
  stair: 'stair',
  barrier: 'barrier',
} as const;

export type SpatialWaypointStatus = typeof SpatialWaypointStatus[keyof typeof SpatialWaypointStatus];


export const SpatialWaypointStatus = {
  clear: 'clear',
  blocked: 'blocked',
  caution: 'caution',
} as const;

export interface SpatialWaypoint {
  id: string;
  label: string;
  type: SpatialWaypointType;
  x: number;
  y: number;
  status: SpatialWaypointStatus;
}

export interface ClockDirections {
  ahead: string;
  right: string;
  left: string;
  behind: string;
}

export type SpatialContextCurrentPoint = {
  x: number;
  y: number;
};

export interface SpatialContext {
  waypoints: SpatialWaypoint[];
  currentPoint: SpatialContextCurrentPoint;
  headingDegrees: number;
  clockDirections: ClockDirections;
}

export interface EnvironmentState {
  id: string;
  observedAt: string;
  pathStatus: EnvironmentStatePathStatus;
  obstacles: string[];
  stairs: boolean;
  elevator: EnvironmentStateElevator;
  detectedObjects: string[];
  signs: string[];
  confidence: number;
  locationLabel: string;
  ocr: OcrResult | null;
  spatialContext: SpatialContext | null;
}

export interface DetectedChange {
  detected: boolean;
  before: string;
  change: string;
  currentState: string;
  impact: string;
  recommendation: string;
}

export type AccessibilityPreferencesResponseLength = typeof AccessibilityPreferencesResponseLength[keyof typeof AccessibilityPreferencesResponseLength];


export const AccessibilityPreferencesResponseLength = {
  brief: 'brief',
  standard: 'standard',
  detailed: 'detailed',
} as const;

export type AccessibilityPreferencesAccessibilityMode = typeof AccessibilityPreferencesAccessibilityMode[keyof typeof AccessibilityPreferencesAccessibilityMode];


export const AccessibilityPreferencesAccessibilityMode = {
  'screen-reader': 'screen-reader',
  'high-contrast': 'high-contrast',
  standard: 'standard',
} as const;

export interface AccessibilityPreferences {
  avoidStairs: boolean;
  preferElevator: boolean;
  avoidCrowds: boolean;
  responseLength: AccessibilityPreferencesResponseLength;
  accessibilityMode: AccessibilityPreferencesAccessibilityMode;
  goal: string;
}

export type PreferencesUpdateResponseLength = typeof PreferencesUpdateResponseLength[keyof typeof PreferencesUpdateResponseLength];


export const PreferencesUpdateResponseLength = {
  brief: 'brief',
  standard: 'standard',
  detailed: 'detailed',
} as const;

export type PreferencesUpdateAccessibilityMode = typeof PreferencesUpdateAccessibilityMode[keyof typeof PreferencesUpdateAccessibilityMode];


export const PreferencesUpdateAccessibilityMode = {
  'screen-reader': 'screen-reader',
  'high-contrast': 'high-contrast',
  standard: 'standard',
} as const;

export interface PreferencesUpdate {
  avoidStairs?: boolean;
  preferElevator?: boolean;
  avoidCrowds?: boolean;
  responseLength?: PreferencesUpdateResponseLength;
  accessibilityMode?: PreferencesUpdateAccessibilityMode;
  goal?: string;
}

export interface EnvironmentDashboard {
  current: EnvironmentState;
  previous: EnvironmentState | null;
  change: DetectedChange;
  preferences: AccessibilityPreferences;
  recommendation: string;
  demoMode: boolean;
  safetyMessage: string;
}

export type ObservationInputPathStatus = typeof ObservationInputPathStatus[keyof typeof ObservationInputPathStatus];


export const ObservationInputPathStatus = {
  clear: 'clear',
  blocked: 'blocked',
} as const;

export type ObservationInputElevator = typeof ObservationInputElevator[keyof typeof ObservationInputElevator];


export const ObservationInputElevator = {
  available: 'available',
  unavailable: 'unavailable',
  unknown: 'unknown',
} as const;

export interface ObservationInput {
  pathStatus: ObservationInputPathStatus;
  obstacles: string[];
  stairs: boolean;
  elevator: ObservationInputElevator;
  detectedObjects: string[];
  signs: string[];
  confidence: number;
  locationLabel: string;
  ocr?: OcrResult | null;
}

export type OcrInputMimeType = typeof OcrInputMimeType[keyof typeof OcrInputMimeType];


export const OcrInputMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface OcrInput {
  /** Base64-encoded image bytes. Used transiently for OCR and never persisted. */
  imageData: string;
  mimeType: OcrInputMimeType;
}

export type AssistantCommandInputSource = typeof AssistantCommandInputSource[keyof typeof AssistantCommandInputSource];


export const AssistantCommandInputSource = {
  voice: 'voice',
  text: 'text',
  camera: 'camera',
} as const;

export interface AssistantCommandInput {
  command: string;
  source: AssistantCommandInputSource;
}

export type AssistantResponseIntent = typeof AssistantResponseIntent[keyof typeof AssistantResponseIntent];


export const AssistantResponseIntent = {
  navigate: 'navigate',
  describe: 'describe',
  changes: 'changes',
  elevator: 'elevator',
  unknown: 'unknown',
} as const;

export interface AssistantResponse {
  transcript: string;
  response: string;
  intent: AssistantResponseIntent;
  confidence: number;
  safetyMessage: string;
}