# lib/api-zod/src/generated/types/assistantCommandInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantCommandInputSource } from './assistantCommandInputSource';

export interface AssistantCommandInput {
  command: string;
  source: AssistantCommandInputSource;
}