# lib/api-zod/src/generated/types/assistantResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantResponseIntent } from './assistantResponseIntent';

export interface AssistantResponse {
  transcript: string;
  response: string;
  intent: AssistantResponseIntent;
  confidence: number;
  safetyMessage: string;
}