# lib/api-zod/src/generated/types/clockDirections.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ClockDirections {
  ahead: string;
  right: string;
  left: string;
  behind: string;
}