# lib/api-zod/src/generated/types/detectedChange.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectedChange {
  detected: boolean;
  before: string;
  change: string;
  currentState: string;
  impact: string;
  recommendation: string;
}