# lib/api-zod/src/generated/types/environmentState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { EnvironmentStateElevator } from './environmentStateElevator';
import type { EnvironmentStatePathStatus } from './environmentStatePathStatus';
import type { OcrResult } from './ocrResult';
import type { SpatialContext } from './spatialContext';

export interface EnvironmentState {
  id: string;
  observedAt: Date;
  pathStatus: EnvironmentStatePathStatus;
  obstacles: string[];
  stairs: boolean;
  elevator: EnvironmentStateElevator;
  detectedObjects: string[];
  signs: string[];
  confidence: number;
  locationLabel: string;
  ocr: OcrResult | null;
  spatialContext: SpatialContext | null;
}