# lib/api-zod/src/generated/types/ocrInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { OcrInputMimeType } from './ocrInputMimeType';

export interface OcrInput {
  /** Base64-encoded image bytes. Used transiently for OCR and never persisted. */
  imageData: string;
  mimeType: OcrInputMimeType;
}