# lib/api-zod/src/generated/types/spatialContext.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ClockDirections } from './clockDirections';
import type { SpatialContextCurrentPoint } from './spatialContextCurrentPoint';
import type { SpatialWaypoint } from './spatialWaypoint';

export interface SpatialContext {
  waypoints: SpatialWaypoint[];
  currentPoint: SpatialContextCurrentPoint;
  headingDegrees: number;
  clockDirections: ClockDirections;
}