# lib/api-zod/src/generated/types/spatialWaypoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ACCESS-X accessibility assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { SpatialWaypointStatus } from './spatialWaypointStatus';
import type { SpatialWaypointType } from './spatialWaypointType';

export interface SpatialWaypoint {
  id: string;
  label: string;
  type: SpatialWaypointType;
  x: number;
  y: number;
  status: SpatialWaypointStatus;
}